backend/requirements-auth.txt
flask-oauthlib>=0.9.6
requests-oauthlib>=1.3.1
PyJWT>=2.6.0
google-auth>=2.16.0